requirements.txt
streamlit>=1.35
langchain-openai>=0.1.0
langchain-core>=0.3.0
pyyaml>=6.0